+++ src/ryQuickPickItemBase.ts
@@ -9,7 +9,11 @@ import * as vscode from 'vscode';
 
 
 
-export class MyQuickPickItemButton implements vscode.QuickInputButton
+/**
+ * QuickPick のボタンを実装したクラス。
+ * ボタンをクリックした時の処理を設定可能。ただし動作させるには onDidTriggerItemButton で execute を呼び出す必要があります。
+ */
+export class RyQuickPickItemButton implements vscode.QuickInputButton
 {
 	private action: (context: vscode.ExtensionContext | undefined) => void | Promise<void>;
 
@@ -23,6 +27,10 @@ export class MyQuickPickItemButton implements vscode.QuickInputButton
 		this.action = action;
 	}
 
+	/**
+	 * action に設定された処理を実行する。
+	 * @param context
+	 */
 	async execute(context: vscode.ExtensionContext | undefined)
 	{
 		await this.action(context);
@@ -38,14 +46,18 @@ export class MyQuickPickItemButton implements vscode.QuickInputButton
 
 
 
-export abstract class MyQuickPickItemBase implements vscode.QuickPickItem
+/**
+ * QuickPickItem を実装した基底クラス。
+ * 選択した時の処理を実行する execute メソッドを定義してある。
+ */
+export abstract class RyQuickPickItemBase implements vscode.QuickPickItem
 {
 	abstract label: string;
 	description?: string;
 	detail?: string;
 	picked?: boolean;
 	alwaysShow?: boolean;
-	buttons?: MyQuickPickItemButton[];
+	buttons?: RyQuickPickItemButton[];
 
 	abstract execute(context: vscode.ExtensionContext | undefined): void | Promise<void>;
 }